ISSUE #366 move files
